OS X fixes

+++ src/os.cpp
@@ -205,7 +205,7 @@ void OS::DispatchKeyboardEvent(const int key, const int scancode, const int acti
 }
 
 void OS::DispatchCharacterEvent(const unsigned int uchar) {
-    KeyboardEvent key_event { uchar, 0, KeyboardEvent::KEY_CHAR, 0 };
+    KeyboardEvent key_event { static_cast<int>(uchar), 0, KeyboardEvent::KEY_CHAR, 0 };
     event::Dispatcher<KeyboardEvent>::GetInstance()->NotifySubscribers(&key_event);
 }
